Make dropdown to bar gap configurable and organized code.
