+++ modules/menus/dashboard/index.ts
@@ -1,4 +1,4 @@
-import DropdownMenu from '../DropdownMenu.js';
+import DropdownMenu from '../shared/dropdown/index.js';
 import { Profile } from './profile/index.js';
 import { Shortcuts } from './shortcuts/index.js';
 import { Controls } from './controls/index.js';
