+++ modules/menus/power/verification.ts
@@ -1,5 +1,5 @@
 import Window from 'types/widgets/window.js';
-import PopupWindow from '../PopupWindow.js';
+import PopupWindow from '../shared/popup/index.js';
 import powermenu from './helpers/actions.js';
 import { Attribute, Child } from 'lib/types/widget.js';
