+++ options.ts
@@ -150,12 +150,13 @@ const options = mkOptions(OPTIONS, {
             layer: opt<WindowLayer>('top'),
             margin_top: opt('0.5em'),
             opacity: opt(100),
-            margin_bottom: opt('0.5em'),
+            margin_bottom: opt('0em'),
             margin_sides: opt('0.5em'),
             border_radius: opt('0.4em'),
             outer_spacing: opt('1.6em'),
             label_spacing: opt('0.5em'),
             transparent: opt(false),
+            dropdownGap: opt('2.9em'),
             background: opt(colors.crust),
             buttons: {
                 style: opt<BarButtonStyles>('default'),
